linted everything, got rid of faulty tests
